% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_chico_shiny}
\alias{run_chico_shiny}
\title{Run the CHICO Shiny Application}
\usage{
run_chico_shiny(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{shiny::shinyApp}}
}
\value{
A Shiny application object
}
\description{
This function launches the CHICO Shiny application.
}
